lessons/02-rag-chatbot/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==2.21.0
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart==0.0.9
pypdf==6.7.1
numpy==1.26.4
